+++ ClassLibraries/Macross.Json.Extensions/Code/System.Text.Json.Serialization/JsonIPAddressConverter.cs
@@ -1,4 +1,10 @@
-﻿using System.Net;
+﻿#if NETSTANDARD2_1_OR_GREATER
+using System.Buffers;
+using System.Net;
+using System.Net.Sockets;
+#else
+using System.Net;
+#endif
 
 using Macross.Json.Extensions;
 
@@ -15,6 +21,15 @@ namespace System.Text.Json.Serialization
 			if (reader.TokenType != JsonTokenType.String)
 				throw ThrowHelper.GenerateJsonException_DeserializeUnableToConvertValue(typeof(IPAddress));
 
+#if NETSTANDARD2_1_OR_GREATER
+			Span<char> charData = stackalloc char[45];
+			int count = Encoding.UTF8.GetChars(
+				reader.HasValueSequence ? reader.ValueSequence.ToArray() : reader.ValueSpan,
+				charData);
+			return !IPAddress.TryParse(charData.Slice(0, count), out IPAddress value)
+				? throw ThrowHelper.GenerateJsonException_DeserializeUnableToConvertValue(typeof(IPAddress))
+				: value;
+#else
 			string value = reader.GetString()!;
 
 			try
@@ -25,13 +40,24 @@ namespace System.Text.Json.Serialization
 			{
 				throw ThrowHelper.GenerateJsonException_DeserializeUnableToConvertValue(typeof(IPAddress), value, ex);
 			}
+#endif
 		}
 
 		/// <inheritdoc/>
-		public override void Write(Utf8JsonWriter writer, IPAddress value, JsonSerializerOptions options) =>
+		public override void Write(Utf8JsonWriter writer, IPAddress value, JsonSerializerOptions options)
+		{
 #pragma warning disable CA1062 // Don't perform checks for performance. Trust our callers will be nice.
+#if NETSTANDARD2_1_OR_GREATER
+			Span<char> data = value.AddressFamily == AddressFamily.InterNetwork
+				? stackalloc char[15]
+				: stackalloc char[45];
+			if (!value.TryFormat(data, out int charsWritten))
+				throw new JsonException($"IPAddress [{value}] could not be written to JSON.");
+			writer.WriteStringValue(data.Slice(0, charsWritten));
+#else
 			writer.WriteStringValue(value.ToString());
+#endif
 #pragma warning restore CA1062 // Validate arguments of public methods
-
+		}
 	}
 }