+++ ClassLibraries/Macross.Json.Extensions/CHANGELOG.md
@@ -27,6 +27,12 @@
 * Added `JsonTypeConverterAdapter` for using `TypeConverter`s with
   System.Text.Json. ([#19](https://github.com/Macross-Software/core/pull/19)) 
 
+* Added `Utf8JsonStreamReader` and improved performance of
+  `JsonIPAddressConverter`, `JsonIPEndPointConverter`, &
+  `JsonTimeSpanConverter` on .NET Standard 2.1+.
+
+* Added `JsonVersionConverter`.
+
 ## 2.0.0
 
 * `JsonMicrosoftDateTimeConverter` & `JsonMicrosoftDateTimeOffsetConverter` now